Add support for key beeps to 2477S switch.
